Create relaymorsevoice
This code tests the relays.  I am using an RGB spotlight with individual input line for each color.  The morse code is used as part of the test.  The voice system "flight" was added to say if a "." or a "-" is sent.
To add flight to the system
sudo apt-get install flite
test with "sudo flite -t "hello I can talk".  For some reason the first word is ignored when I run "flite"